requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
owlready2==0.47
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
pyswip==0.3.2
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.1
scipy==1.15.0
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
tabulate==0.9.0
threadpoolctl==3.5.0
tzdata==2024.2
xgboost==2.1.3
